Update index.js
Modified termination type to "TADO_MODE" instead of "MANUAL". This allows the user to temporarily adjust the temperature/power of the tado without disrupting the tado's Smart Schedule. Submitting "MANUAL" causes the tado to ignore the Smart Schedule and Proximity Detection, and the temperature/power can then only be controlled manually. If users want their tado to always be on manual mode, they can simply disable the smart schedule and proximity detection in the tado app.